+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 4.3.3
++ [ ] DoubleSlider: new components
 
 ### 4.3.2
 + [ ] Table: new attributes `data-head`, `data-body`, `data-static`